Include agenda in pre tag with calendar.
Combined the header and footer into a single "body" variable.  The pre tag now
includes both of these blocks of text.  The content of the script, however, is
not being included now, since it can not be easily eascaped (since the email
is in HTML).  It could be escaped with some simple string replacement, but
that should probably be best done in a Perl script or something.