Create remind file for Google Calendars.
Take a list of links to Google Calendar iCal files in an input file (for now,
the file must be named "calendar_list" in the same directory).  For each link,
download the file to the /tmp directory, convert it to remind format using the
ruby script ical2rem-cli, and then concatenate it with the previous calendars.