Added full paths to commands and files.
The shell that cron runs scripts in runs them from the user's home directory,
not the directory with the script, so relative paths must account for that.
This was causing the script to not be able to find the calendar_list file.

The script run by cron also does not have the same PATH as the normal user.
This was preventing the script from finding the ical2rem-cli and org2remind.pl
files.  For now, their full paths have been hard-coded in.  Ultimately, they
(and the location of the calendar_list) file can be set in configuration
variables.