Change wget to not print any output.
The -nv flag still printed output for each file that was downloaded, which is
not needed in the normal case (plus it puts the URLs for the calendars in more
locations).  This would not print error information, but wget does not provide
any additional output if an attempt is made to download a URL that does not
exist.